Program/requirements.txt
flet
google-generative-ai
google-auth
requests
httpx
